image-pipeline/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
Pillow==11.0.0
numpy==2.1.0
PyMuPDF==1.25.0
olefile==0.47
supabase==2.11.0
python-dotenv==1.0.1
